Check error logs before testing redirect
We want to send more detailed error logs in our failure emails. Currently, whichever error is caught first is sent in the failure email. By catching our cloudwatch error logs first, rather than our redirect errors, our failure emails will be more descriptive.

Note that this does not change the requirements for a failure email to be sent. Either a cloudwatch error log or redirect error will still cause a failure email to be sent.

Also note that a cloudwatch error log is likely to occur alongside a redirect error - because the redirects are caught first, this is why there have been so few emails with cloudwatch error logs.